WIP: wrong!! WB_MakeWaveBuilderWave: Fix pulse starting time calcuation
Since 0e52306f (WB_MakeWaveBuilderWave: Factor out epochID filling into
WB_UpdateEpochID, 2017-10-17) the pulse starting times for all but the
first epoch are wrong.

This is because the accumulatedDuration was only calculated if the
parameter updateEpochIDWave is true, and that is only the case when
creating the stimset in the wavebuilder for display.

In a future commit we will introduce some logic to force recreation of
all stimsets.