SweepBrowser/DataBrowser: Make all axes scaling options available to both
Before the merge of the sweep browser and databrowser code bases in
80053ff6 (introduce BrowserSettings panel with tabcontrol, 2017-08-29)
we had different options for axes scaling available in both.

The merge kept that distinction.

But we can easily make all three options available in both of them. This
has the additional advantage that we don't need the finalUpdateHook in
the PostPlotSettings anymore, thus making the graph update code easier
to grasp and change.